refactor: enum id resolver location
