+++ src/main/java/cl/bennu/commons/enums/QueueEnum.java
@@ -1,12 +1,10 @@
 package cl.bennu.commons.enums;
 
 import cl.bennu.commons.enums.base.BaseEnum;
+import cl.bennu.commons.enums.base.EnumIdResolver;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
-
-import java.util.Arrays;
-import java.util.Map;
 
 @Getter
 @JsonFormat(shape = JsonFormat.Shape.OBJECT)
